+++ profiling_tools/clibs/calcBin.h
@@ -1,8 +1,8 @@
-int binary_srch(long *arr, int l, int r, int x);
+int binary_srch(long long *arr, int l, int r, int x);
 void calcHistogram(int *histogram_counter, int histogram_size,
-		long *histogram_index,
-		double *histogram_util, 
-		long *critical_points, int critical_points_size,
-		long *location_index, int location_size,
+		long long *histogram_index,
+		double *histogram_util,
+		long long *critical_points, int critical_points_size,
+		long long *location_index, int location_size,
 		int *location_counter,
 		double *location_util);
